+++ scaffolds/js-component/template/src/components/{{name}}.js
@@ -1,5 +1,0 @@
-import {{classifiedName}}Template from './{{name}}.sht';
-
-export default class {{classifiedName}} extends {{classifiedName}}Template {
-
-}